requirements.txt
fastapi==0.115.6
loguru==0.7.2
openai==2.29.0
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.12.5
httpx==0.28.1
uvicorn==0.34.0
